+++ packages/office-addin-manifest/src/manifestInfo.ts
@@ -1,12 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { usageDataObject } from "./defaults";
-import { ManifestHandler } from "./manifestHandler/manifestHandler";
-import { getManifestHandler } from "./manifestHandler/getManifestHandler";
-import * as util from "util";
-import * as fs from "fs";
-
 export class DefaultSettings {
   public sourceLocation?: string;
   public requestedWidth?: string;
@@ -31,58 +25,4 @@ export class ManifestInfo {
   public version?: string;
 
   public defaultSettings?: DefaultSettings;
-}
-
-export namespace OfficeAddinManifest {
-  export async function modifyManifestFile(
-    manifestPath: string,
-    guid?: string,
-    displayName?: string
-  ): Promise<ManifestInfo> {
-    let manifestData: ManifestInfo = {};
-    if (manifestPath) {
-      if (guid === undefined && displayName === undefined) {
-        throw new Error("You need to specify something to change in the manifest.");
-      } else {
-        try {
-          const manifestHandler: ManifestHandler = getManifestHandler(manifestPath);
-          const fileData: string = await readFromManifestFile(manifestPath);
-
-          manifestData = await manifestHandler.modifyManifest(manifestPath, fileData, guid, displayName);
-          await manifestHandler.writeManifestData(manifestPath, manifestData);
-          let output = await readManifestFile(manifestPath);
-          usageDataObject.reportSuccess("modifyManifestFile()");
-          return output;
-        } catch (err: any) {
-          usageDataObject.reportException("modifyManifestFile()", err);
-          throw err;
-        }
-      }
-    } else {
-      throw new Error(`Please provide the path to the manifest file.`);
-    }
-  }
-
-  export async function readManifestFile(manifestPath: string): Promise<ManifestInfo> {
-    if (manifestPath) {
-      const manifestHandler: ManifestHandler = getManifestHandler(manifestPath);
-      const fileData: string = await readFromManifestFile(manifestPath);
-
-      const manifest: ManifestInfo = await manifestHandler.parseManifest(manifestPath, fileData);
-      return manifest;
-    } else {
-      throw new Error(`Please provide the path to the manifest file.`);
-    }
-  }
-}
-
-async function readFromManifestFile(manifestPath: string): Promise<string> {
-  try {
-    const fileData: string = await util.promisify(fs.readFile)(manifestPath, {
-      encoding: "utf8",
-    });
-    return fileData;
-  } catch (err) {
-    throw new Error(`Unable to read data for manifest file: ${manifestPath}. \n${err}`);
-  }
 }
